requirements.txt
streamlit
tensorflow==2.15.0
keras==2.15.0
transformers==4.36.2
torch
numpy
pillow
huggingface_hub